refactor: split hook-module into hook-esm and hook-cjs
There is no reason to keep these two hooks inside a single component. 
All the other hooks have their dedicated component. Also remove deadcode 
for transformSourceDefault.
