+++ components/.test.list
@@ -41,7 +41,8 @@
 - hook-fixture
 - * hook-apply
   * hook-group
-  * hook-module
+  * hook-cjs
+  * hook-esm
   * hook-query
   * - hook-http
     - * hook-http-client
