+++ components/hook-cjs/node/.build.yml
@@ -1,0 +1,4 @@
+branches: [test, node]
+dependencies:
+  - util
+  - agent
