+++ components/hook-cjs/void/index.mjs
@@ -1,0 +1,12 @@
+export default (dependencies) => {
+  const {
+    util: { noop },
+    expect: { expect },
+  } = dependencies;
+  return {
+    hook: (agent, { hooks: { cjs } }) => {
+      expect(!cjs, "expected configuration to disable cjs module hook");
+    },
+    unhook: noop,
+  };
+};
