+++ components/hook-cjs/void/index.test.mjs
@@ -3,16 +3,16 @@ import {
   buildTestDependenciesAsync,
   buildTestComponentAsync,
 } from "../../build.mjs";
-import HookNativeModule from "./index.mjs";
+import HookModule from "./index.mjs";
 
 const dependencies = await buildTestDependenciesAsync(import.meta.url);
 const { testHookAsync } = await buildTestComponentAsync("hook-fixture");
-const component = HookNativeModule(dependencies);
+const component = HookModule(dependencies);
 assertDeepEqual(
   await testHookAsync(
     component,
-    { hooks: { esm: false, cjs: false } },
-    async () => {},
+    { hooks: { cjs: false } },
+    async (state) => null,
   ),
   { sources: [], events: [] },
 );
