+++ components/hook-esm/node/.eslintrc.yml
@@ -1,0 +1,2 @@
+env:
+  node: true
