+++ components/hook-esm/void/.build.yml
@@ -1,0 +1,2 @@
+branches: [browser]
+dependencies: [util, expect]
