--- components/hook-module/node/.test.list
@@ -1,3 +1,0 @@
-- * common.mjs
-  * native.mjs
-- index.mjs
