+++ components/hook/default/index.mjs
@@ -1,17 +1,16 @@
 const { fromEntries } = Object;
 
 export default (dependencies) => {
-  const { "hook-module": transformSourceDefault } = dependencies;
   const names = [
     "hook-apply",
     "hook-group",
-    "hook-module",
+    "hook-esm",
+    "hook-cjs",
     "hook-query",
     "hook-http-client",
     "hook-http-server",
   ];
   return {
-    transformSourceDefault,
     hook: (agent, configuration) =>
       fromEntries(
         names.map((name) => [
